feat(2147): add view model
