+++ test/presentation/user_action/user_action_create_view_model_test.dart
@@ -1,4 +1,3 @@
-import 'package:clock/clock.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:pass_emploi_app/features/user_action/create/user_action_create_actions.dart';
 import 'package:pass_emploi_app/features/user_action/create/user_action_create_state.dart';
@@ -10,7 +9,6 @@ import 'package:pass_emploi_app/redux/app_state.dart';
 import 'package:redux/redux.dart';
 
 import '../../doubles/spies.dart';
-import '../../dsl/app_state_dsl.dart';
 
 void main() {
   test("create when state is loading should set display state to loading", () {
@@ -62,64 +60,6 @@ void main() {
 
     // Then
     expect(viewModel.displayState, isA<DismissWithFailure>());
-  });
-
-  group('isRappelActive', () {
-    final store = givenState().store();
-    final viewModel = UserActionCreateViewModel.create(store);
-
-    test('when dateEcheance is null should return false', () {
-      // Given
-      const DateTime? dateEcheance = null;
-
-      // When
-      final bool result = viewModel.isRappelActive(dateEcheance);
-
-      // Then
-      expect(result, isFalse);
-    });
-
-    test('when dateEcheance is in past should return false', () {
-      final today = DateTime(2022, 1, 2);
-      withClock(Clock.fixed(today), () {
-        // Given
-        final dateEcheance = DateTime(2022, 1, 1);
-
-        // When
-        final bool result = viewModel.isRappelActive(dateEcheance);
-
-        // Then
-        expect(result, isFalse);
-      });
-    });
-
-    test('when dateEcheance is before day + 3 should return false', () {
-      final today = DateTime(2022, 1, 2);
-      withClock(Clock.fixed(today), () {
-        // Given
-        final dateEcheance = DateTime(2022, 1, 5);
-
-        // When
-        final bool result = viewModel.isRappelActive(dateEcheance);
-
-        // Then
-        expect(result, isFalse);
-      });
-    });
-
-    test('when dateEcheance is after day + 3 should return true', () {
-      final today = DateTime(2022, 1, 2);
-      withClock(Clock.fixed(today), () {
-        // Given
-        final dateEcheance = DateTime(2022, 1, 6);
-
-        // When
-        final bool result = viewModel.isRappelActive(dateEcheance);
-
-        // Then
-        expect(result, isTrue);
-      });
-    });
   });
 
   test('createUserAction should dispatch CreateUserAction', () {
